refactor(ui): move network fields to component
Make a reusable component for the network fields.

Fixes: #2203.
