--- ui/src/app/machines/views/MachineDetails/MachineNetwork/AddInterface/AddInterfaceFormFields/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddInterfaceFormFields";
