+++ ui/src/app/machines/views/MachineDetails/MachineNetwork/NetworkFields/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NetworkFields";
